// client/src/actions/types.tsx
export enum TypeKeys {
  GET_ERRORS = "GET_ERRORS",
  SET_CURRENT_USER = "SET_CURRENT_USER",
  GET_PROFILE = "GET_PROFILE",
  PROFILE_LOADING = "PROFILE_LOADING",
  PROFILE_NOT_FOUND = "PROFILE_NOT_FOUND",
  CLEAR_CURRENT_PROFILE = "CLEAR_CURRENT_PROFILE",
  GET_PROFILES = "GET_PROFILES",
  POST_LOADING = "POST_LOADING",
  GET_POSTS = "GET_POSTS",
  GET_POST = "GET_POST",
  ADD_POST = "ADD_POST",
  DELETE_POST = "DELETE_POST",
  CLEAR_ERRORS = "CLEAR_ERRORS"
}